// src/api/openapi/models/UpdateAboutUsRequest.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type UpdateAboutUsRequest = {
    about_us?: string;
    city_commitment?: string;
    president_message?: string;
    banner?: Blob;
    city_image?: Blob;
    president_image?: Blob;
    qcu_logo?: Blob;
    agrihub_user_logo?: Blob;
};